backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
opencv-python-headless==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
lxml==4.9.3
aiofiles==23.2.1
